Problem: reading large images (in my case, it was a 40792x20480 RGBA PNG one) causes a crash, when the product row-number * width-step overflows a 32-bit integer (#3091)
Solution: change the type of the width step to be `size_t` (and not `long`)

+++ dlib/array2d/array2d_kernel_abstract.h
@@ -242,7 +242,7 @@ namespace dlib
                 - returns #*this
         !*/
 
-        long width_step (
+        size_t width_step (
         ) const;
         /*!
             ensures
